pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "documind"
version = "0.1.0"
description = "Multi-Agent Document Analysis System for Enterprise"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    # LLM & AI Frameworks
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-community>=0.3.0",
    "openai>=1.50.0",
    "anthropic>=0.35.0",
    "transformers>=4.45.0",
    "torch>=2.4.0",
    "peft>=0.13.0",
    "sentence-transformers>=3.0.0",
    
    # Vector Store & Embeddings
    "qdrant-client>=1.11.0",
    "cohere>=5.10.0",
    
    # API Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    
    # Document Processing
    "pypdf>=4.3.0",
    "python-docx>=1.1.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.4.0",
    
    # Storage & Caching
    "boto3>=1.35.0",
    "redis>=5.1.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    
    # Monitoring & Observability
    "prometheus-client>=0.21.0",
    "langsmith>=0.1.0",
    "structlog>=24.4.0",
    
    # Report Generation
    "reportlab>=4.2.0",
    "weasyprint>=62.0",
    "jinja2>=3.1.0",
    
    # Utilities
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ragas>=0.2.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
    "bandit>=1.7.0",
    "ipython>=8.27.0",
]

[project.scripts]
documind = "documind.main:run"

[project.urls]
Homepage = "https://github.com/yourusername/documind"
Repository = "https://github.com/yourusername/documind"
Documentation = "https://github.com/yourusername/documind/docs"

[tool.hatch.build.targets.wheel]
packages = ["src/documind"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["documind"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "langgraph.*",
    "qdrant_client.*",
    "cohere.*",
    "transformers.*",
    "torch.*",
    "peft.*",
    "reportlab.*",
    "weasyprint.*",
    "pytesseract.*",
    "docx.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src/documind"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
